% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOpenscmRunner-package.R
\docType{package}
\name{rOpenscmRunner-package}
\alias{rOpenscmRunner}
\alias{rOpenscmRunner-package}
\title{rOpenscmRunner: Run different simple climate models from R using a unified interface}
\description{
Using openscm-runner, you can run different simple climate models using a unified API. It supports emissions-driven runs only. rOpenscmRunner is a wrapper to easily use openscm-runner from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/rOpenscmRunner}
}

}
\author{
\strong{Maintainer}: Mika Pflüger \email{mika.pflueger@pik-potsdam.de}

}
