% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  climateModelsConfigs,
  scenarios,
  outputVariables = list("Surface Temperature"),
  outConfig = NULL,
  returnRaw = FALSE
)
}
\arguments{
\item{climateModelsConfigs}{Named list where the names are climate models and the corresponding list member is a
configuration or list of configurations to use for this model. Use NULL to denote the default configuration with
no changes. Climate model names supported: CiceroSCM, FaIR, and MAGICC7. The supported configuration settings
depend on the used climate model, please refer to their respective documentation.}

\item{scenarios}{DataFrame containing one or multiple emissions scenarios to simulate. The data shoud be in
IAMC format, i.e. a wide data frame with the years as columns. Index columns that are additionally required are
"model", "scenario", "region", "variable", and "unit". The "variable" must also follow the IAMC format naming
conventions, e.g. "Emissions|CO2" for the 'total CO2 emissions (not including CCS)'.}

\item{outputVariables}{A list of variables to include in the output.}

\item{outConfig}{Named list where the names are climate models and the corresponding list member is a list of
configuration values to include in the output in the metadata. Optional, default: don't include input variables in
the output metadata.}

\item{returnRaw}{Boolean to control the return type. By default (returnRaw = FALSE), run() returns a named list
result, where result$df is a data frame which contains the result in the same format as the input scenarios, and
result$metadata is a named list of metadata generated during the run. If returnRaw is TRUE, run() returns a python
scmdata.ScmRun object.}
}
\description{
Run one or multiple climate models over one or multiple emissions scenarios.
}
\details{
Some simple climate models need certain information to run. Use the setup function to supply this information
before attempting to use run.
}
\examples{
\dontrun{
# create very minimal emissions scenario.
df <- data.frame(
 model = c("rand", "rand"),
 scenario = c("weirdEMI", "weirdEMI"),
 region = c("World", "World"),
 variable = c("Emissions|CO2", "Emissions|CH4"),
 unit = c("Mt CO2 / yr", "Mt CH4 / yr"),
 "2015" = c(9., 12.),
 "2020" = c(10., 11.),
 check.names = FALSE)

# simulate the scenario using MAGICC7 with default settings.
run(climateModelsConfigs = list("MAGICC7" = NULL), scenarios = df)

# simulate the scenario using MAGICC7 with default settings but "somesetting" set to "12"
run(climateModelsConfigs = list("MAGICC7" = list("somesetting" = "12")), scenarios = df)

# simulate the scenario with MAGICC7 and FaIR, each with two different sets of configurations
# where in MAGICC7 we change "somesetting" and in FaIR we change "fairsetting".
# Also include the changed configuration settings in the output.
run(climateModelsConfigs = list("MAGICC7" = list(list("somesetting" = "12"),
                                                 list("somesetting" = "13")),
                                "FaIR" = list(list("fairsetting" = "slr"),
                                              list("fairsetting" = "noslr"))),
    scenarios = df,
    outConfig = list("MAGICC7" = list("somesetting"), "FaIR" = list("fairsetting")))
}
}
\author{
Mika Pflüger
}
