% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{setup}
\usage{
setup(
  magiccExecutable7 = NULL,
  magiccWorkerNumber = NULL,
  magiccWorkerRootDir = NULL
)
}
\arguments{
\item{magiccExecutable7}{The file path of a MAGICC version 7 executable file (with corresponding config file
structure around it). You have to supply this if you intend to run MAGICC version 7. Instead of using this
function to supply it, you can also set the environment variable MAGICC_EXECUTABLE_7 before starting your R session.}

\item{magiccWorkerNumber}{The number of processes which should be started when running MAGICC. By default, as many
processes are started as there are processors in your system. Instead of using this function to supply the setting,
you can also set the environment variable MAGICC_WORKER_NUMBER before starting your R session.}

\item{magiccWorkerRootDir}{The path to a folder where the temporary directories to run MAGICC will be created.
By default, a standard location is chosen depending on your operating system. Instead of using this function to
supply the setting, you can also set the environment variable MAGICC_WORKER_ROOT_DIR before starting your R session.
If you do supply a non-standard directory, you have to create and clean it up yourself. If you use the standard
settings, everything happens automatically.}
}
\description{
Prepare rOpenscmRunner to run climate models.
}
\details{
Some simple climate models need certain information to run. This function provides a unified
interface to supply this information.
}
\examples{
setup(magiccExecutable7 = "/path/to/magicc/bin/magicc", magiccWorkerNumber = 4)
}
\author{
Mika Pflüger
}
