% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  climateModelsConfigs,
  scenarios,
  outputVariables = list("Surface Temperature"),
  outConfig = NULL
)
}
\arguments{
\item{climateModelsConfigs}{Named list where the names are climate models and the corresponding list member is a
configuration or list of configurations to use for this model. Use null to denote the default configuration with
no changes. Climate model names supported: CiceroSCM, FaIR, and MAGICC7. The supported configuration settings
depend on the used climate model, please refer to their respective documentation.}

\item{scenarios}{DataFrame containing one or multiple emissions scenarios to simulate.}

\item{outputVariables}{A list of variables to include in the output. Optional, default: list("Surface Temperature")}

\item{outConfig}{Named list where the names are climate models and the corresponding list member is a list of
configuration values to include in the output in the metadata. Optional, default: don't include input variables in
the output metadata.}
}
\description{
Run one or multiple climate models over one or multiple emissions scenarios.
}
\details{
Some simple climate models need certain information to run. Use the setup function to supply this information
before attempting to use run.
}
\examples{
\dontrun{
run(climateModelsConfigs = list("MAGICC7" = NULL), scenarios = ...)

run(climateModelsConfigs = list("MAGICC7" = list("somesetting" = "12")), scenarios = ...)

run(climateModelsConfigs = list("MAGICC7" = list(list("somesetting" = "12"),
                                                 list("somesetting" = "13")),
                                "FaIR" = list(list("fairsetting" = "slr"),
                                              list("fairsetting" = "noslr"))),
    scenarios = ...,
    outConfig = list("MAGICC7" = list("somesetting"), "FaIR" = list("fairsetting")))
}
}
\author{
Mika Pflüger
}
